Enforce workspace-root single-namespace contract: move normalizeFormattedOutput into Plugin namespace
The @gml-modules/plugin workspace root (plugin/index.ts) was exporting
`normalizeFormattedOutput` as a flat named export alongside the `Plugin`
namespace, violating the target-state architectural contract (§2.1) that
each workspace root must export exactly one named namespace.

- Add `normalizeFormattedOutput` to GmlPlugin type with ownership TSDoc
- Add `normalizeFormattedOutput` to the Plugin object in plugin-entry.ts
- Remove flat export from plugin/index.ts (workspace root now exports only Plugin)
- Remove flat re-export from plugin/src/index.ts (internal barrel)
- Update CLI to access normalizeFormattedOutput via moduleValue.Plugin
- Update test imports to use direct relative paths per internal-import rule
