+++ src/plugin/index.ts
@@ -1,2 +1,1 @@
 export { Plugin } from "./src/index.js";
-export { normalizeFormattedOutput } from "./src/printer/normalize-formatted-output.js";
