+++ src/plugin/src/index.ts
@@ -1,2 +1,1 @@
 export { defaultOptions, languages, pluginOptions as options, parsers, Plugin, printers } from "./plugin-entry.js";
-export { normalizeFormattedOutput } from "./printer/normalize-formatted-output.js";
